refactored playground (including signing model activities), and games, and deprecated files, into packages for easier management
