+++ app/src/main/AndroidManifest.xml
@@ -33,18 +33,19 @@
         <activity android:name=".features.auth.SignupPage" />
         <activity android:name=".features.tabs.HomePage" />
         <activity android:name=".features.auth.TestPage" />
-        <activity android:name=".signrecognition.MainActivity3" />
-        <activity android:name=".signrecognition.videorecognition.RecordVideoActivity" />
-        <activity android:name=".signrecognition.liverecognition.LiveSignRecognitionActivity" />
-        <activity android:name=".signrecognition.videorecognition.SignRecognitionResultActivity" />
-        <activity android:name=".LoadImageActivity" />
+        <activity android:name=".deprecated.MainActivity3" />
+        <activity android:name=".features.tabs.playground.videorecognition.RecordVideoActivity" />
+        <activity android:name=".features.tabs.playground.liverecognition.LiveSignRecognitionActivity" />
+        <activity android:name=".features.tabs.playground.videorecognition.SignRecognitionResultActivity" />
+        <activity android:name=".deprecated.LoadImageActivity" />
         <activity android:name=".features.tabs.discussion.DiscussionPage" />
-        <activity android:name=".features.games.MatchingGameActivity" />
+        <activity android:name=".features.tabs.playground.PlaygroundActivity" />
+        <activity android:name=".games.MatchingGameActivity" />
         <activity android:name=".features.games.SelectingGameActivity" />
         <activity android:name=".features.games.FillBlankGameActivity" />
         <activity android:name=".features.games.IdentifyGameActivity" />
-        <activity android:name=".features.games.SigningGameActivity" />
-        <activity android:name=".features.games.SpellWordGameActivity" />
+        <activity android:name=".games.SigningGameActivity" />
+        <activity android:name=".games.SpellWordGameActivity" />
         <activity android:name=".features.activitycenter.ActivityCenter" />
 
 <!--        <activity android:name=".features.tabs.resources.ResourcesPage" />-->
