+++ app/src/main/java/com/example/signifybasic/features/activitycenter/ActivityCenter.kt
@@ -3,40 +3,27 @@ package com.example.signifybasic.features.activitycenter
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
 import android.util.Log
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ProgressBar
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
-import com.example.signifybasic.features.games.FillBlankGameActivity
-import com.example.signifybasic.features.games.FillBlankGameData
-import com.example.signifybasic.features.games.FillBlankOption
 import com.example.signifybasic.R
-import com.example.signifybasic.features.games.SelectingGameActivity
-import com.example.signifybasic.features.games.SelectingGameData
-import com.example.signifybasic.features.games.GameModule
-import com.example.signifybasic.features.games.IdentifyGameActivity
-import com.example.signifybasic.features.games.IdentifyGameData
-import com.example.signifybasic.features.games.IdentifyOption
-import com.example.signifybasic.features.games.MatchingGameActivity
-import com.example.signifybasic.features.games.MatchingItem
+import com.example.signifybasic.games.GameModule
 import com.example.signifybasic.features.utility.applyHighContrastToAllViews
 import com.example.signifybasic.features.utility.applyTextSizeToAllTextViews
 import com.example.signifybasic.features.utility.isHighContrastEnabled
 import com.example.signifybasic.getStarted
 import com.google.android.material.appbar.MaterialToolbar
-import com.example.signifybasic.features.games.GameRouter
-import com.example.signifybasic.features.games.GameSequenceManager
-import com.example.signifybasic.features.games.GameStep
-import com.example.signifybasic.features.games.ModuleManager
+import com.example.signifybasic.games.GameRouter
+import com.example.signifybasic.games.GameSequenceManager
+import com.example.signifybasic.games.ModuleManager
 import com.example.signifybasic.database.DBHelper
-import com.example.signifybasic.features.games.ModuleManager.currentStepIndex
+import com.example.signifybasic.games.ModuleManager.currentStepIndex
 
 
 class ActivityCenter : AppCompatActivity() {
