+++ app/src/main/java/com/example/signifybasic/features/tabs/home/HomeFragment.kt
@@ -12,13 +12,12 @@ import androidx.cardview.widget.CardView
 import com.example.signifybasic.R
 import com.example.signifybasic.features.activitycenter.ActivityCenter
 import com.example.signifybasic.features.tabs.discussion.DiscussionFragment
+import com.example.signifybasic.features.tabs.playground.PlaygroundActivity
 import com.example.signifybasic.features.tabs.resources.ResourcesFragment
 import com.example.signifybasic.features.tabs.settings.SettingsFragment
 import com.example.signifybasic.features.utility.applyHighContrastToAllViews
 import com.example.signifybasic.features.utility.applyTextSizeToAllTextViews
 import com.example.signifybasic.features.utility.isHighContrastEnabled
-import com.example.signifybasic.signrecognition.liverecognition.LiveSignRecognitionActivity
-import com.example.signifybasic.signrecognition.MainActivity3
 
 class HomeFragment : Fragment() {
 
@@ -57,11 +56,11 @@ class HomeFragment : Fragment() {
         }
 
         cardDiscussion.setOnClickListener {
-            startActivity(Intent(requireContext(), MainActivity3::class.java))
-//            requireActivity().supportFragmentManager.beginTransaction()
-//                .replace(R.id.fragment_container, DiscussionFragment())
-//                .addToBackStack(null)
-//                .commit()
+//            startActivity(Intent(requireContext(), RecordVideoActivity::class.java))
+            requireActivity().supportFragmentManager.beginTransaction()
+                .replace(R.id.fragment_container, DiscussionFragment())
+                .addToBackStack(null)
+                .commit()
         }
 
         cardResources.setOnClickListener{
@@ -77,7 +76,7 @@ class HomeFragment : Fragment() {
 
         cardPlayground.setOnClickListener {
             //playground
-            startActivity(Intent(requireContext(), LiveSignRecognitionActivity::class.java))
+            startActivity(Intent(requireContext(), PlaygroundActivity::class.java))
         }
 
         return view
