+++ app/src/main/java/com/example/signifybasic/features/tabs/playground/videorecognition/ModelApiService.kt
@@ -1,4 +1,4 @@
-package com.example.signifybasic.signrecognition.videorecognition
+package com.example.signifybasic.features.tabs.playground.videorecognition
 
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
