+++ app/src/main/java/com/example/signifybasic/games/GameModule.kt
@@ -1,6 +1,4 @@
-package com.example.signifybasic.features.games
-
-import com.example.signifybasic.features.games.GameStep
+package com.example.signifybasic.games
 
 data class GameModule(
     val moduleId: String,
