+++ app/src/main/java/com/example/signifybasic/games/GameRouter.kt
@@ -1,13 +1,11 @@
-package com.example.signifybasic.features.games
+package com.example.signifybasic.games
 
 import android.content.Context
 import android.content.Intent
 import android.widget.Toast
 import com.example.signifybasic.features.games.FillBlankGameActivity
-import com.example.signifybasic.features.games.SelectingGameActivity
 import com.example.signifybasic.features.games.IdentifyGameActivity
-import com.example.signifybasic.features.games.MatchingGameActivity
-import com.example.signifybasic.features.games.SigningGameActivity
+import com.example.signifybasic.features.games.SelectingGameActivity
 
 object GameRouter {
     fun routeToGame(context: Context, stepIndex: Int) {
