+++ app/src/main/res/layout/activity_record_video.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:background="@drawable/login_background"
-    tools:context=".signrecognition.videorecognition.RecordVideoActivity">
+    tools:context=".features.tabs.playground.videorecognition.RecordVideoActivity">
 
     <!-- loading animation -->
     <ProgressBar
